refactor: change user validation
